+++ pkg/_macros/lib/src/executor/augmentation_library.dart
@@ -202,6 +202,7 @@ class _Builder {
     }
     final Set<Identifier> mergedAugmentedTypes = {
       ...mergedEntryResults.keys,
+      ...mergedExtendsResults.keys,
       ...mergedInterfaceResults.keys,
       ...mergedMixinResults.keys,
       ...mergedTypeResults.keys,